backend/requirements.txt
fastapi[all]
uvicorn
pymongo
pydantic
motor
python-jose[cryptography]
passlib[bcrypt]
email-validator

